+++ app/src/main/res/layout/fragment_contact_detail.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentaion.ContactDetailFragment">
+    tools:context=".presentation.ContactDetailFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
